Fixed SDLColor (full class reflection was drawn but left incomplete).
git-svn-id: svn://falconpl.org/falcon/modules/sdl/trunk@62 08e93009-9248-0410-8e9b-f56be5c5ec23

+++ src/sdl_events_ext.cpp
@@ -1321,7 +1321,7 @@ FALCON_FUNC sdl_JoystickEventState( VMachine *vm )
       return;
    }
 
-   vm->retval( (int64) ::SDL_JoystickEventState( i_code->asInteger() ) );
+   vm->retval( (int64) ::SDL_JoystickEventState( (int) i_code->asInteger() ) );
 }
 
 /*#
